move Individual mode to own file and spellcorrect
